+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+
+-   Fixed the ForeignKeyConstraint error when refreshing a directory that contains a file with a linked Download (@Vinnybod)
 -   Downgraded bcrypt to version 4.0.1 to resolve issue in passlib (@Cx01N)
 
 ## [5.9.1] - 2024-01-25